Fix: Relocation login modal
